+++ lib/src/string_values.dart
@@ -1,66 +1,66 @@
 /// Empty string: ''
-const emptyString = string.empty;
+const String emptyString = string.empty;
 
 /// chars used by base64 encoding
-const base62chars = string.base62digits;
+const String base62chars = string.base62digits;
 
 /// chars used by base64 encoding
-const base64chars = string.base64digits;
+const String base64chars = string.base64digits;
 
 // python inspired
 
 /// The lowercase letters 'abcdefghijklmnopqrstuvwxyz'.
 /// This value is not locale-dependent and will not change.
-const asciiLowercase = string.asciiLowercase;
+const String asciiLowercase = string.asciiLowercase;
 
 /// The uppercase letters 'ABCDEFGHIJKLMNOPQRSTUVWXYZ'.
 /// This value is not locale-dependent and will not change.
-const asciiUppercase = string.asciiUppercase;
+const String asciiUppercase = string.asciiUppercase;
 
 /// The concatenation of the [asciiLowercase] and [asciiUppercase] constants.
 /// This value is not locale-dependent.
-const asciiLetters = string.asciiLetters;
+const String asciiLetters = string.asciiLetters;
 
 /// The string '0123456789'.
-const digits = string.digits;
+const String digits = string.digits;
 
 /// The string '0123456789abcdefABCDEF'.
-const hexdigits = string.hexdigits;
+const String hexdigits = string.hexdigits;
 
 /// The string '01234567'.
-const octdigits = string.octdigits;
+const String octdigits = string.octdigits;
 
 /// String of ASCII characters which are considered punctuation characters in
 /// the C locale: !"#$%&'()*+,-./:;<=>?@[\]^_`{|}~.
-const punctuation = string.punctuation;
+const String punctuation = string.punctuation;
 
 /// A string containing all ASCII characters that are considered whitespace.
 /// This includes the characters space, tab, linefeed, return, formfeed,
 /// and vertical tab.
-const whitespace = string.whitespace;
+const String whitespace = string.whitespace;
 
 /// a string containing all Unicode characters that are considered whitespace.
 ///
 /// taken from https://en.wikipedia.org/wiki/Whitespace_character
-const allWhitespace = string.allWhitespace;
+const String allWhitespace = string.allWhitespace;
 
 /// all Unicode space characters with White_Space=yes
 ///
 /// taken from https://en.wikipedia.org/wiki/Whitespace_character
-const visibleWhitespace = string.visibleWhitespace;
+const String visibleWhitespace = string.visibleWhitespace;
 
 /// all Unicode space characters with White_Space=no
 ///
 /// taken from https://en.wikipedia.org/wiki/Whitespace_character
-const invisibleWhitespace = string.invisibleWhitespace;
+const String invisibleWhitespace = string.invisibleWhitespace;
 
 /// String of ASCII characters which are considered printable.
 /// This is a combination of digits, ascii_letters, punctuation, and whitespace.
-const printable = string.printable;
+const String printable = string.printable;
 
 /// String of ASCII characters which are considered printable except spaces.
 /// This is a combination of digits, ascii_letters, and punctuation.
-const printableNoSpaces = string.printableNoSpaces;
+const String printableNoSpaces = string.printableNoSpaces;
 
 /// static class to be used like in python
 /// The constants defined as in python https://docs.python.org/3/library/string.html
@@ -69,57 +69,57 @@ class string {
   const string._();
 
   /// Empty string: ''
-  static const empty = '';
+  static const String empty = '';
 
   /// chars used by base64 encoding
-  static const base62digits = digits + asciiUppercase + asciiLowercase;
+  static const String base62digits = digits + asciiUppercase + asciiLowercase;
 
   /// chars used by base64 encoding
-  static const base64digits = '$base62digits+/';
+  static const String base64digits = '$base62digits+/';
 
   // python inspired
 
   /// The lowercase letters 'abcdefghijklmnopqrstuvwxyz'.
   /// This value is not locale-dependent and will not change.
-  static const asciiLowercase = 'abcdefghijklmnopqrstuvwxyz';
+  static const String asciiLowercase = 'abcdefghijklmnopqrstuvwxyz';
 
   /// The uppercase letters 'ABCDEFGHIJKLMNOPQRSTUVWXYZ'.
   /// This value is not locale-dependent and will not change.
-  static const asciiUppercase = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ';
+  static const String asciiUppercase = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ';
 
   /// The concatenation of the [asciiLowercase] and [asciiUppercase]  constants.
   /// This value is not locale-dependent.
-  static const asciiLetters =
+  static const String asciiLetters =
       'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ';
 
   /// The string '0123456789'.
-  static const digits = '0123456789';
+  static const String digits = '0123456789';
 
   /// The string '0123456789abcdefABCDEF'.
-  static const hexdigits = '0123456789abcdefABCDEF';
+  static const String hexdigits = '0123456789abcdefABCDEF';
 
   /// The string '01234567'.
-  static const octdigits = '01234567';
+  static const String octdigits = '01234567';
 
   /// String of ASCII characters which are considered punctuation characters in
   /// the C locale: !"#$%&'()*+,-./:;<=>?@[\]^_`{|}~.
-  static const punctuation = r'''!"#$%&'()*+,-./:;<=>?@[\]^_`{|}~''';
+  static const String punctuation = r'''!"#$%&'()*+,-./:;<=>?@[\]^_`{|}~''';
 
   /// A string containing all ASCII characters that are considered whitespace.
   /// This includes the characters space, tab, linefeed, return, formfeed,
   /// and vertical tab.
-  static const whitespace = ' \u{0009}\n\r\u{000C}\u{2B7F}';
+  static const String whitespace = ' \u{0009}\n\r\u{000C}\u{2B7F}';
 
   /// a string containing all Unicode characters that are considered whitespace.
   ///
   /// taken from https://en.wikipedia.org/wiki/Whitespace_character
-  static const allWhitespace =
+  static const String allWhitespace =
       string.visibleWhitespace + string.invisibleWhitespace;
 
   /// all Unicode space characters with White_Space=yes
   ///
   /// taken from https://en.wikipedia.org/wiki/Whitespace_character
-  static const visibleWhitespace = '\u{0009}'
+  static const String visibleWhitespace = '\u{0009}'
       '\u{000A}'
       '\u{000B}'
       '\u{000C}'
@@ -155,7 +155,7 @@ class string {
   /// all Unicode space characters with White_Space=no
   ///
   /// taken from https://en.wikipedia.org/wiki/Whitespace_character
-  static const invisibleWhitespace = '\u{180E}'
+  static const String invisibleWhitespace = '\u{180E}'
       '\u{200B}'
       '\u{200C}'
       '\u{200D}'
@@ -167,10 +167,10 @@ class string {
 
   /// String of ASCII characters which are considered printable except spaces.
   /// This is a combination of digits, ascii_letters, and punctuation.
-  static const printableNoSpaces = digits + asciiLetters + punctuation;
+  static const String printableNoSpaces = digits + asciiLetters + punctuation;
 
   /// String of ASCII characters which are considered printable.
   /// This is a combination of digits, ascii_letters, punctuation,
   /// and whitespace.
-  static const printable = printableNoSpaces + whitespace;
+  static const String printable = printableNoSpaces + whitespace;
 }